messaging-app/backend/run.py
#!/usr/bin/env python3
"""
Messaging App Backend Startup Script
"""

import uvicorn
import os
from app.config import settings

if __name__ == "__main__":
    # Load environment variables
    from dotenv import load_dotenv
    load_dotenv()
    
    # Run the application
    uvicorn.run(
        "app.main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info",
        access_log=True
    )
