messaging-app/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
websockets==15.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
motor==3.7.1
pymongo==4.14.1
beanie==2.0.0
redis==6.4.0
celery==5.5.3
cryptography==45.0.7
python-socketio==5.13.0
pydantic==2.11.7
python-dotenv==1.0.0
email-validator==2.1.1
